get_ua/requirements.txt
pycryptodome==3.19.0
bce-python-sdk==0.8.97
cup==3.2.32
unrar==0.4
PyMySQL==1.1.0
Pillow==10.1.0
opencv-python==4.8.1.78
lxml==4.9.1
python-dotenv==1.0.0
requests==2.31.0
XlsxWriter==3.1.9
xlrd==2.0.1
chardet==5.2.0
urllib3==2.1.0
python-dateutil==2.8.2
zhconv==1.4.3
numpy
pyinstaller
sqlalchemy
mysql-connector-python
